Cambiar gestión de cantidades de List a Map en Pedido
- Cambiar la propiedad 'cantidades' de List<Integer> a Map<Integer, Integer> en Pedido.
- Añadir constructor de 4 parámetros y el método getCliente() en Pedido.
- Refactorizar métodos internos (calcularTotal, addProducto, etc.) para operar con las claves del mapa usando producto.getId().
- Actualizar Tienda.realizarVenta() para acceder a las cantidades por ID de producto en lugar de por índice.
- Adaptar llamadas internas en PedidoTest a la nueva estructura de datos, manteniendo intactas las aserciones.

+++ entornos/src/main/java/com/entornos/ProductoFisico.java
@@ -26,9 +26,9 @@ public class ProductoFisico extends Producto{
      * @param peso El peso del paquete en kg.
      * @param destino El destino de envío.
      */
-    public ProductoFisico(int id, String nombre, float precioBase, float peso, String destino){
+    public ProductoFisico(int id, String nombre, float precioBase, double peso, String destino){
         super(id, nombre, precioBase);
-        this.peso = peso;
+        this.peso = (float) peso;
         this.destino = destino;
     }
 
@@ -39,9 +39,9 @@ public class ProductoFisico extends Producto{
      * @param precioBase El precio base del producto.
      * @param peso El peso del paquete en kg.
      */
-    public ProductoFisico(int id, String nombre, float precioBase, float peso){
+    public ProductoFisico(int id, String nombre, float precioBase, double peso){
         super(id, nombre, precioBase);
-        this.peso = peso;
+        this.peso = (float) peso;
         this.destino = "No especificado";
     }
